#81 Makes displaying language of a user to follow his configuration
