+++ src/frontend/js/esn/app.run.js
@@ -1,23 +1,9 @@
-(function(angular) {
-  'use strict';
+'use strict';
 
-  angular.module('esnApp')
-    // don't remove $state from here or ui-router won't route...
-    .run(function(session, ioConnectionManager, $state) { // eslint-disable-line
-      session.ready.then(function() {
-        ioConnectionManager.connect();
-      });
-    })
-    .run(settingLanguage);
-
-  function settingLanguage($cookies, $translate, esnConfig, ESN_I18N_DEFAULT_LOCALE) {
-    esnConfig('core.language')
-      .then(function(language) {
-        $cookies.locale = language;
-        $translate.use(language);
-      })
-      .catch(function() {
-        $cookies.locale = ESN_I18N_DEFAULT_LOCALE;
-      });
-  }
-})(angular);
+angular.module('esnApp')
+  // don't remove $state from here or ui-router won't route...
+  .run(function(session, ioConnectionManager, $state) { // eslint-disable-line
+    session.ready.then(function() {
+      ioConnectionManager.connect();
+    });
+  });
