+++ src/frontend/js/modules/i18n/i18n.run.js
@@ -1,10 +1,23 @@
+'use strict';
+
 require('./i18n.service.js');
+require('./i18n.constants');
 
-(function(angular) {
-  'use strict';
+angular.module('esn.i18n')
+  .run(setMomentLocale)
+  .run(settingLanguage);
 
-  angular.module('esn.i18n')
-    .run(function(amMoment, esnI18nService) {
-      amMoment.changeLocale(esnI18nService.getLocale());
+function setMomentLocale(amMoment, esnI18nService) {
+  amMoment.changeLocale(esnI18nService.getLocale());
+}
+
+function settingLanguage($cookies, $translate, esnConfig, ESN_I18N_DEFAULT_LOCALE) {
+  esnConfig('core.language')
+    .then(function(language) {
+      $cookies.locale = language;
+      $translate.use(language);
+    })
+    .catch(function() {
+      $cookies.locale = ESN_I18N_DEFAULT_LOCALE;
     });
-})(angular);
+}
